+++ src/main/kotlin/io/runescript/plugin/ide/sdk/RsSdkSetupValidator.kt
@@ -11,7 +11,6 @@ import com.intellij.ui.EditorNotificationPanel
 import io.runescript.plugin.ide.RsBundle
 import io.runescript.plugin.ide.filetypes.RsFileType
 import io.runescript.plugin.lang.RuneScript
-import io.runescript.plugin.oplang.filetypes.RsOpFileType
 
 
 class RsSdkSetupValidator : ProjectSdkSetupValidator {
@@ -19,7 +18,7 @@ class RsSdkSetupValidator : ProjectSdkSetupValidator {
     override fun isApplicableFor(project: Project, file: VirtualFile): Boolean {
         val fileType = file.fileType
         when (fileType) {
-            is RsFileType, RsOpFileType -> {
+            is RsFileType -> {
                 val psiFile = PsiManager.getInstance(project).findFile(file) ?: return false
                 return psiFile.language.isKindOf(RuneScript)
             }