+++ src/main/kotlin/io/runescript/plugin/oplang/lexer/RsOpLexer.kt
@@ -1,6 +1,0 @@
-package io.runescript.plugin.oplang.lexer
-
-import com.intellij.lexer.FlexAdapter
-import io.runescript.plugin.lang.lexer.RsLexerInfo
-
-class RsOpLexer(lexerInfo: RsLexerInfo) : FlexAdapter(_RsOpLexer(null, lexerInfo))