+++ src/main/kotlin/io/runescript/plugin/oplang/psi/stub/RsOpReturnListStub.kt
@@ -1,9 +1,0 @@
-package io.runescript.plugin.oplang.psi.stub
-
-import com.intellij.psi.stubs.IStubElementType
-import com.intellij.psi.stubs.StubBase
-import com.intellij.psi.stubs.StubElement
-import io.runescript.plugin.oplang.psi.RsOpReturnList
-
-class RsOpReturnListStub(parent: StubElement<*>?, elementType: IStubElementType<*, *>)
-    : StubBase<RsOpReturnList>(parent, elementType)